backend/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Deprecated==1.2.14
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-heroku==0.3.1
django-ipware==6.0.3
djangorestframework==3.14.0
ecdsa==0.18.0
gunicorn==21.2.0
idna==3.6
jwcrypto==1.5.1
jwt==1.3.1
mysqlclient==2.2.1
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-ipware==2.0.1
python-jose==3.3.0
python-jwt==4.1.0
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
whitenoise==6.6.0
wrapt==1.16.0
